Add files via upload
Project-2 is a twitter mining, where we choose a random twitter user and find his friends and followers. Later we find the reciprocal friends and select the top 5 reciprocal friends by their their followers count recursively until we gather a minimum of 100 nodes. Now, using these 100 nodes we create a social network graph using netwrokx package. Finally we save the graph as an image and calculate the network size, diameter and average distance for the graph  and save them in a file. Saving graph is done using matplotlib package